Small refactor in mass calibration module
