+++ src/main/java/io/github/mzmine/modules/dataprocessing/masscalibration/errormodeling/errortypes/PpmError.java
@@ -16,30 +16,25 @@
  * USA
  */
 
-package io.github.mzmine.modules.dataprocessing.masscalibration.errormodeling;
+package io.github.mzmine.modules.dataprocessing.masscalibration.errormodeling.errortypes;
 
-public interface ErrorType {
-  /**
-   * Calculate error given measured and actual values
-   *
-   * @param measured
-   * @param actual
-   * @return
-   */
-  double calculateError(double measured, double actual);
+/**
+ * Ppm error type,
+ * parts per million is just fractional multiplied by 10^6
+ */
+public class PpmError implements ErrorType {
+  public double calculateError(double measured, double actual) {
+    double fractionalError = (measured - actual) / actual;
+    return fractionalError * 1_000_000;
+  }
 
-  /**
-   * Calibrate a value against an error
-   * given measured and its error towards actual should return back actual value
-   * ie, the methods should be related as:
-   * <p>
-   * error = calculateError(measured, actual)
-   * calibrated = calibrateAgainstError(measured, error)
-   * calibrated == actual
-   *
-   * @param value
-   * @param error
-   * @return
-   */
-  double calibrateAgainstError(double value, double error);
+  public double calibrateAgainstError(double value, double error) {
+    double fractionalError = error / 1_000_000;
+    return value / (1 + fractionalError);
+  }
+
+  @Override
+  public String toString() {
+    return "PPM error";
+  }
 }
